@@ -1,21 +1,20 @@
 {
-  "name": "cleanfolio",
-  "version": "0.1.0",
+  "name": "sajan-web-portfolio",
+  "version": "1.0.0",
   "private": true,
-  "homepage": "https://rjshkhr.github.io/cleanfolio",
+  "homepage": "sajanghuman.com",
   "dependencies": {
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/icons": "^4.11.2",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.0.0",
-    "@testing-library/user-event": "^13.2.1",
-    "gh-pages": "^3.2.3",
-    "prop-types": "^15.7.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "4.0.3",
+    "@emotion/react": "^11.11.0",
+    "@emotion/styled": "^11.11.0",
+    "@mui/material": "^5.15.0",
+    "@mui/icons-material": "^5.15.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-scripts": "^5.0.1",
+    "gh-pages": "^6.3.0",
+    "prop-types": "^15.8.1",
     "uniqid": "^5.4.0",
-    "web-vitals": "^2.1.0"
+    "web-vitals": "^3.3.0"
   },
   "scripts": {
     "start": "react-scripts start --openssl-legacy-provider",
