@@ -1,11 +1,13 @@
-import { render } from 'react-dom'
+import { createRoot } from 'react-dom/client'
 import App from './App'
 import { ThemeProvider } from './contexts/theme'
 import './index.css'
 
-render(
+const container = document.getElementById('root')
+const root = createRoot(container)
+
+root.render(
   <ThemeProvider>
     <App />
-  </ThemeProvider>,
-  document.getElementById('root')
+  </ThemeProvider>
 )
