@@ -1,71 +1,112 @@
 const header = {
-  // all the properties are optional - can be left empty or deleted
-  homepage: 'https://rjshkhr.github.io/cleanfolio',
-  title: 'JS.',
+  homepage: 'sajanghuman.com',
+  title: 'Portfolio',
 }
 
 const about = {
-  // all the properties are optional - can be left empty or deleted
-  name: 'John Smith',
-  role: 'Front End Engineer',
+  name: 'Sajan',
+  role: 'Full Stack Developer',
   description:
-    'Adipisicing sit fugit ullam unde aliquid sequi Facilis soluta facilis perspiciatis corporis nulla aspernatur. Autem eligendi rerum delectus modi quisquam? Illo ut quasi nemo ipsa cumque perspiciatis! Maiores minima consectetur.',
-  resume: 'https://example.com',
+    'Greetings! 👋 I am Sajan, a passionate and motivated web developer currently honing my skills at Bluewave Labs. 🍁',
   social: {
-    linkedin: 'https://linkedin.com',
-    github: 'https://github.com',
+    linkedin: 'https://www.linkedin.com/in/sajanghuman',
+    github: 'https://github.com/SajanGhuman',
   },
 }
 
 const projects = [
-  // projects can be added an removed
-  // if there are no projects, Projects section won't show up
   {
-    name: 'Project 1',
-    description:
-      'Amet asperiores et impedit aliquam consectetur? Voluptates sed a nulla ipsa officia et esse aliquam',
-    stack: ['SASS', 'TypeScript', 'React'],
-    sourceCode: 'https://github.com',
-    livePreview: 'https://github.com',
+    name: 'Datahall',
+    description: 'An open source DocSend alternative.',
+    stack: [
+      'NextJS',
+      'Typescript',
+      'Tailwind',
+      'Docker',
+      'Supabase',
+      'PrismaORM',
+    ],
+    sourceCode: 'https://github.com/SajanGhuman/Minecraft-3JS',
+    livePreview: 'https://minecraft.sajanghuman.com/',
+    blog: 'https://www.linkedin.com/posts/sajanghuman_bluewavedatahall-mvplaunch-datamanagement-activity-7287899878557458432-Xn4X?utm_source=share&utm_medium=member_desktop&rcm=ACoAAEZACZgBKhkysouDqbrDUPe0VCxiN7m78io',
   },
   {
-    name: 'Project 2',
+    name: 'Padraig',
     description:
-      'Amet asperiores et impedit aliquam consectetur? Voluptates sed a nulla ipsa officia et esse aliquam',
-    stack: ['SASS', 'TypeScript', 'React'],
-    sourceCode: 'https://github.com',
-    livePreview: 'https://github.com',
+      'Learning Management System for scheduling online meetings and delivering training courses to company employees',
+    stack: ['React', 'Tailwind', 'PostgresSQL', 'Express', 'NodeJS'],
+    sourceCode: '',
+    livePreview: 'https://www.padraig.ca',
   },
   {
-    name: 'Project 3',
-    description:
-      'Amet asperiores et impedit aliquam consectetur? Voluptates sed a nulla ipsa officia et esse aliquam',
-    stack: ['SASS', 'TypeScript', 'React'],
-    sourceCode: 'https://github.com',
-    livePreview: 'https://github.com',
+    name: 'Minecraft Clone',
+    description: 'Minecraft clone made in pure Typescript and WebGL.',
+    stack: ['WebGL', 'TypeScript'],
+    sourceCode: 'https://github.com/SajanGhuman/Minecraft-3JS',
+    livePreview: 'https://minecraft.sajanghuman.com/',
+    blog: 'https://www.linkedin.com/posts/sajanghuman_typescript-threejs-gamedevelopment-activity-7228171617271496704-l0-s?utm_source=share&utm_medium=member_desktop&rcm=ACoAAEZACZgBKhkysouDqbrDUPe0VCxiN7m78io',
+  },
+  {
+    name: 'Cube Space',
+    description: 'Learn rubciks cube algorithms interactively.',
+    stack: ['TypeScript', 'React', 'Tailwind CSS'],
+    sourceCode: 'https://github.com/SajanGhuman/Cube-Space',
+    livePreview: 'https://cubespace.sajanghuman.com/',
+  },
+  {
+    name: 'Naruto Sensei',
+    description: 'Talk to a 3D AI-powered chatgpt about naruto anime series.',
+    stack: ['Three JS', 'TypeScript', 'React', 'ChatGPT API'],
+    sourceCode: 'https://github.com/SajanGhuman/ask_naruto_sensei',
+    livePreview: 'https://naruto.sajanghuman.com/',
+    blog: 'https://www.linkedin.com/posts/sajanghuman_threejs-ai-javascript-activity-7229977306688020481-kdgq?utm_source=share&utm_medium=member_desktop&rcm=ACoAAEZACZgBKhkysouDqbrDUPe0VCxiN7m78io',
+  },
+  {
+    name: 'Winnpeg Parks',
+    description: 'Simple API to search nearby parks in Winnipeg.',
+    stack: ['Javascript', 'CSS', 'Winnipeg Open Data API'],
+    sourceCode: 'https://github.com/SajanGhuman/winnipeg-parks',
+    livePreview: 'https://parks.sajanghuman.com/',
+  },
+  {
+    name: 'Tic Tac Toe',
+    description: 'Classic Tic-Tac-Toe game that we all love.',
+    stack: ['React', 'CSS'],
+    sourceCode: 'https://github.com/SajanGhuman/Tic-Tac-Toe',
+    livePreview: 'https://tictac.sajanghuman.com/',
   },
 ]
 
 const skills = [
-  // skills can be added or removed
-  // if there are no skills, Skills section won't show up
   'HTML',
   'CSS',
   'JavaScript',
   'TypeScript',
   'React',
-  'Redux',
-  'SASS',
+  'NextJS',
+  'Tailwind',
   'Material UI',
+  'SASS',
+  'Redux',
+  'Zustand',
+  'NodeJS',
+  'Express',
+  'RESTful API',
+  'PostgresSQL',
+  'MongoDB',
+  'MySQL',
+  'NoSQL',
   'Git',
+  'Git Lab',
   'CI/CD',
-  'Jest',
-  'Enzyme',
+  'Docker',
+  'Linux',
+  'Neo Vim',
+  'Scrum/Kanban',
 ]
 
 const contact = {
-  // email is optional - if left empty Contact section won't show up
-  email: 'johnsmith@mail.com',
+  email: 'sajanghuman18@gmail.com',
 }
 
 export { header, about, projects, skills, contact }
